Spike - Does the Fastly decacher know if has decached an article?
If so, then it will be able to decide whether or not to ping Facebook Newstab.

The challenge is we seem to receive a RetrievableContent payload which does not contain the content type or the webUrl.

To resolve this we can either:
- bulk up the RetrievableContent event to contain content type and webUrl.
- call back to CAPI from the Fastly Lambda to obtain this information.

The later seems to be what RetrievableContent implies.