Clear live blog ajax files for updates which arrive as EventType.Content and EventType.RetrievableUpdate.
To my untrained eye EventType.Content and EventType.RetrievableUpdate seem to by different physical representations of the same real world update.

The service which published these events (Crier) appears to sometime convert an EventType.Update to an EventType.RetrievableUpdate. The criteria for doing this is the size of the update.
So we might have had a situation where the decache behaviours of large and small live blogs differed subtly.

This commit naively assumes that EventType.Content and EventType.RetrievableUpdate are different implementations of the same thing and should have the same effect.