Update jackson-core to remove snyk vulnerability
Snyk has reported [a high-severity vulnerability in
jackson-core](https://security.snyk.io/vuln/SNYK-JAVA-COMFASTERXMLJACKSONCORE-7569538).
While the maintainers of jackson and play are sceptical about its
severity, it’s easier for us to manually update it than figure out
whether we’re actually vulnerable.

Snyk reports any version from 2.15.0 forward as having the fix, so I’ve
added a dependency on jackson-core:2.15.4 to evict the lower version in
our transitive dependencies (which comes from various aws dependencies
and logstash-logback-encoder).

I haven’t upgraded to the newest possible version for now, as we can
probably remove this explicit dependency when we update our other
dependencies, and they may not move to the newest version of
jackson (which might produce an odd downgrade for us).