Describe deployment process in readme
Before now, the deployment process was not described in the readme – it
was not immediately clear that one must manually update the
cloudformation, or run a manual deployment after merging to main.

Having this explained in the readme would have saved me a small bit of
working out recently, so this commit does that.