Set Riff Raff’s allowedStages to just PROD
Since we don’t currently have a CODE stack for this, it would be good to
prevent people from attempting to deploy to CODE, as a number of us have
recently done. Setting this option should mean that Riff Raff doesn’t
give people a choice to deploy to CODE, indicating to them that it’s
unsupported before they do it.